pyproject.toml
[project]
name = "jobrecommendationsystem"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apify-client>=2.4.0",
    "langchain>=1.2.4",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.0",
    "mcp[cli]>=1.25.0",
    "pymupdf>=1.26.7",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.53.0",
]
